feature: add functionality to explay array with different fudging conditions
the fudge section in the configuration file determines whether to use: value, mean, median, gaussian or noise option. The noise option will create an spectrum with an average and mean of 1 and the noise according to the sds provided. The idea is to use a std that resembles the meanSnr. The average and mean of one is ideal since each spec in the train set is normalized by the median. This is the equivalent to the hide_color=None in LimeImageExplainer where it takes the mean value per super pixel per channel as the fudge, just that in this case we incorporate the noise factor of our data ;)