feature: implement Fudge.gaussians
User can add a gaussian per segment. The sign of the gaussian's amplitude is selected randomly. The standar deviation of the gaussian is controlled by the number of standar deviations the user wnats to fit in the segment. For instance if this value is 8, then, the boundaries of the segment will add the gaussian at mu+4*sigma. Large values are encouraged to keep continuity at the boundaries of the segments