feature: generate plost of clustered anomalies
The plots contain the median and mean of both the explanation weights and of the top 10k anomalous spectra. the clustering was performed with Kmeans and the final number of cluster was decided according to the elbow heuristics. The total number of clusters was 7. The clustering was performed on the explanation weights of the whole wavelength range